+++ app/src/main/java/src/Pieces/Horse.java
@@ -1,6 +1,6 @@
 package src.Pieces;
 
-import src.PieceMovements.KnightMove;
+import src.PiecesMoves.KnightMove;
 import src.Position;
 
 public class Horse extends Piece{
